// frontend/src/api/model/consultationResponse.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * 텍스트 벡터 검색 시스템
 * 교사-학부모 정기상담 내용 벡터 검색 시스템
 * OpenAPI spec version: 1.0.0
 */

export interface ConsultationResponse {
  /**
   * 상담 내용
   * @minLength 1
   */
  text: string;
  id: number;
  created_at: string;
  updated_at: string;
}